requirements.txt
accelerate==1.10.1
aiofiles==25.1.0
aiohappyeyeballs==2.6.1
aiohttp==3.13.1
aiosignal==1.4.0
annotated-types==0.7.0
anyio==4.11.0
async-timeout==4.0.3
attrs==25.4.0
backoff==2.2.1
beautifulsoup4==4.14.2
certifi==2025.10.5
cffi==2.0.0
charset-normalizer==3.4.4
click==8.3.0
colorama==0.4.6
coloredlogs==15.0.1
cryptography==46.0.3
dataclasses-json==0.6.7
emoji==2.15.0
exceptiongroup==1.3.0
faiss-cpu==1.12.0
fastapi==0.119.0
filelock==3.20.0
filetype==1.2.0
FlashRank==0.2.10
flatbuffers==25.9.23
frozenlist==1.8.0
fsspec==2025.9.0
greenlet==3.2.4
h11==0.16.0
hf-xet==1.1.10
html5lib==1.1
httpcore==1.0.9
httpx==0.28.1
httpx-sse==0.4.3
huggingface-hub==0.35.3
humanfriendly==10.0
idna==3.11
Jinja2==3.1.6
joblib==1.5.2
jsonpatch==1.33
jsonpointer==3.0.0
langchain==1.0.0
langchain-classic==1.0.0
langchain-community==0.3.31
langchain-core==1.0.0
langchain-ollama==1.0.0
langchain-text-splitters==0.3.11
langdetect==1.0.9
langgraph==1.0.0
langgraph-checkpoint==2.1.2
langgraph-prebuilt==1.0.0
langgraph-sdk==0.2.9
langsmith==0.4.37
lxml==6.0.2
MarkupSafe==2.1.5
marshmallow==3.26.1
mpmath==1.3.0
multidict==6.7.0
mypy_extensions==1.1.0
networkx==3.3
nltk==3.9.2
numpy==1.26.4
olefile==0.47
ollama==0.6.0
onnxruntime==1.23.1
orjson==3.11.3
ormsgpack==1.11.0
packaging==23.2
pillow==11.3.0
propcache==0.4.1
protobuf==6.33.0
psutil==7.1.0
pycparser==2.23
pydantic==2.12.3
pydantic-settings==2.11.0
pydantic_core==2.41.4
pypdf==6.1.1
pyreadline3==3.5.4
python-docx==1.2.0
python-dotenv==1.1.1
python-iso639==2025.2.18
python-magic==0.4.27
python-oxmsg==0.0.2
PyYAML==6.0.3
rank-bm25==0.2.2
RapidFuzz==3.14.1
regex==2025.9.18
requests==2.32.5
requests-toolbelt==1.0.0
safetensors==0.6.2
scikit-learn==1.7.2
scipy==1.15.3
sentence-transformers==5.1.1
six==1.17.0
sniffio==1.3.1
soupsieve==2.8
SQLAlchemy==2.0.44
starlette==0.48.0
sympy==1.14.0
tenacity==8.5.0
threadpoolctl==3.6.0
tokenizers==0.22.1
torch==2.9.0+cpu
torchaudio==2.9.0+cpu
torchvision==0.24.0+cpu
tqdm==4.67.1
transformers==4.57.1
typing-inspect==0.9.0
typing-inspection==0.4.2
typing_extensions==4.15.0
unstructured==0.18.15
unstructured-client==0.42.3
urllib3==2.5.0
uvicorn==0.37.0
watchdog==6.0.0
webencodings==0.5.1
wrapt==1.17.3
xxhash==3.6.0
yarl==1.22.0
zstandard==0.25.0
